+++ examples/celery_app.rs
@@ -2,7 +2,7 @@
 
 use async_trait::async_trait;
 use celery::error::TaskError;
-use celery::task::TaskSendOptions;
+use celery::task::{Task, TaskSendOptions};
 use env_logger::Env;
 use exitfailure::ExitFailure;
 use structopt::StructOpt;
@@ -28,6 +28,12 @@ fn buggy_task() {
 async fn long_running_task(secs: Option<u64>) {
     let secs = secs.unwrap_or(10);
     time::delay_for(Duration::from_secs(secs)).await;
+}
+
+// Demonstrates a task that is bound to the task instance, i.e. runs as an instance method.
+#[celery::task(bind = true)]
+fn bound_task(task: &Self) -> Option<u32> {
+    task.timeout()
 }
 
 #[derive(Debug, StructOpt)]